+++ EVTCAnalytics/GameData/Encounters/Encounter.cs
@@ -105,6 +105,9 @@ namespace GW2Scratch.EVTCAnalytics.GameData.Encounters
 		Boneskinner = 40004,
 		WhisperOfJormag = 40005,
 		VariniaStormsounder = 40006, // Last enemy in Strike Mission: Cold War
+		
+		// Story
+		Mordremoth = 50001,
 
 		// Other
 		WorldVersusWorld = 100000,