+++ EVTCAnalytics/GameData/SpeciesIds.cs
@@ -113,5 +113,8 @@ namespace GW2Scratch.EVTCAnalytics.GameData
 		public const int Boneskinner = 22521;
 		public const int WhisperOfJormag = 22711;
 		public const int VariniaStormsounder = 22836;
+		
+		// Story
+		public const int HeartsAndMindsMordremoth = 15884;
 	}
 }